Move schema generation from test
