+++ src/lib.rs
@@ -1,0 +1,3 @@
+mod main;
+
+pub use main::{Message, Notification, Request, Response, WebViewOptions};
